+++ amiga_ros2_bridge/launch/twist_joy.launch.py
@@ -1,0 +1,14 @@
+from launch import LaunchDescription
+from launch_ros.actions import Node
+
+
+def generate_launch_description():
+    return LaunchDescription(
+        [
+            Node(
+                package="amiga_ros2_bridge",
+                executable="twist_joy",
+                name="twist_joy",
+            )
+        ]
+    )
